+++ app/src/main/java/com/example/myplayer/viewmodel/PlayerActivityViewModel.kt
@@ -1,13 +1,10 @@
 package com.example.myplayer.viewmodel
 
-import android.graphics.drawable.Drawable
 import androidx.hilt.lifecycle.ViewModelInject
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import com.example.myplayer.core.base.BaseViewModel
 
 class PlayerActivityViewModel @ViewModelInject constructor(
 
-): BaseViewModel() {
+) : BaseViewModel() {
 
 }
